requirements.txt
blinker==1.4
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cssselect2==0.3.0
cycler==0.10.0
defusedxml==0.6.0
dnspython==2.0.0
dominate==2.5.2
EasyProcess==0.3
email-validator==1.1.1
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WeasyPrint==0.6
Flask-WTF==0.14.3
fpdf==1.7.2
html5lib==1.1
idna==2.10
install==1.3.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.19.2
openpyxl==3.0.5
pdfkit==0.6.1
Pillow==7.2.0
pycparser==2.20
pyparsing==2.4.7
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
PyVirtualDisplay==1.3.2
six==1.15.0
SQLAlchemy==1.3.19
tinycss2==1.0.2
visitor==0.1.3
WeasyPrint==51
webencodings==0.5.1
Werkzeug==1.0.1
wkhtmltopdf==0.2
WTForms==2.3.3
XlsxWriter==1.3.3
xlwt==1.3.0
